Added that files should always be downloaded and new field to LocalFile.
Removed code in files-view.coffee that stopped a file from being
downloaded if it had previously been downloaded and the remote file had
the same mtime as the previously downloaded one.

Added a dtime field to LocalFile to allow user to easier distinguish
between previosly downloaded versions of a file.
This is displayed in the "Show Open Files" dialog.

Added check on deserializing that removes LocalFiles if the underlying
file on localhost has been deleted. This can occur if temporary files is
deleted.